+++ packages/atomic/cypress/e2e/facets/numeric-facet/numeric-facet.cypress.ts
@@ -1128,7 +1128,10 @@ describe('Numeric Facet V1 Test Suites', () => {
       }
 
       describe('verify rendering', () => {
-        before(setupSelectedNumericFacetValue);
+        before(() => {
+          setupSelectedNumericFacetValue();
+          cy.wait(TestFixture.interceptAliases.Search);
+        });
         CommonAssertions.assertAccessibility(breadboxComponent);
         BreadboxAssertions.assertDisplayBreadcrumb(true);
         BreadboxAssertions.assertDisplayBreadcrumbClearAllButton(true);
@@ -1148,7 +1151,10 @@ describe('Numeric Facet V1 Test Suites', () => {
         }
 
         describe('verify rendering', () => {
-          before(setupDeselectNumericFacetValue);
+          before(() => {
+            setupDeselectNumericFacetValue();
+            cy.wait(TestFixture.interceptAliases.Search);
+          });
           BreadboxAssertions.assertDisplayBreadcrumb(false);
         });
 
@@ -1190,7 +1196,10 @@ describe('Numeric Facet V1 Test Suites', () => {
       }
 
       describe('verify rendering', () => {
-        before(setupSelectedMulitpleNumericFacetValuess);
+        before(() => {
+          setupSelectedMulitpleNumericFacetValuess();
+          cy.wait(TestFixture.interceptAliases.Search);
+        });
         CommonAssertions.assertAccessibility(breadboxComponent);
         BreadboxAssertions.assertDisplayBreadcrumb(true);
         BreadboxAssertions.assertDisplayBreadcrumbClearAllButton(true);