+++ packages/atomic/src/components/recommendations/atomic-recs-result-template/atomic-recs-result-template.tsx
@@ -2,6 +2,7 @@ import {Component, Element, Prop, Method, State} from '@stencil/core';
 import {RecsResultTemplateCondition, RecsResultTemplate} from '..';
 import {MapProp} from '../../../utils/props-utils';
 import {
+  makeDefinedConditions,
   makeMatchConditions,
   ResultTemplateCommon,
 } from '../../common/result-templates/result-template-common';
@@ -74,6 +75,7 @@ export class AtomicRecsResultTemplate {
   }
 
   public componentWillLoad() {
+    this.conditions = makeDefinedConditions(this.ifDefined, this.ifNotDefined);
     this.resultTemplateCommon.matchConditions = makeMatchConditions(
       this.mustMatch,
       this.mustNotMatch