+++ packages/quantic/force-app/main/default/lwc/quanticUtils/__tests__/quanticUtils.test.js
@@ -1,4 +1,4 @@
-import {I18nUtils} from "c/quanticUtils";
+import {I18nUtils, copyToClipboard} from 'c/quanticUtils';
 
 describe('c/quanticUtils', () => {
   describe('I18nUtils', () => {
@@ -6,7 +6,9 @@ describe('c/quanticUtils', () => {
       const text = 'sample text';
       const startTag = '<test-start-tag>';
       const endTag = '<test-end-tag>';
-      expect(I18nUtils.getTextWithDecorator(text, startTag, endTag)).toBe(`${startTag}${text}${endTag}`);
+      expect(I18nUtils.getTextWithDecorator(text, startTag, endTag)).toBe(
+        `${startTag}${text}${endTag}`
+      );
     });
 
     it('getTextBold should return text wrapped in bold tags', () => {
@@ -18,36 +20,48 @@ describe('c/quanticUtils', () => {
       const testLabelName = 'thisLabelName';
       const testLabelNamePlural = `${testLabelName}_plural`;
       const testLabelNameZero = `${testLabelName}_zero`;
-      
+
       it('should return plural variant if count not equal to 1', () => {
-        expect(I18nUtils.getLabelNameWithCount(testLabelName, 99)).toBe(testLabelNamePlural);
+        expect(I18nUtils.getLabelNameWithCount(testLabelName, 99)).toBe(
+          testLabelNamePlural
+        );
       });
 
       it('should return plural variant if count is fraction of 1', () => {
-        expect(I18nUtils.getLabelNameWithCount(testLabelName, 0.5)).toBe(testLabelNamePlural);
+        expect(I18nUtils.getLabelNameWithCount(testLabelName, 0.5)).toBe(
+          testLabelNamePlural
+        );
       });
 
       it('should return zero variant if count is 0', () => {
-        expect(I18nUtils.getLabelNameWithCount(testLabelName, 0)).toBe(testLabelNameZero);
+        expect(I18nUtils.getLabelNameWithCount(testLabelName, 0)).toBe(
+          testLabelNameZero
+        );
       });
 
       it('should return singular variant if count is equal to 1', () => {
-        expect(I18nUtils.getLabelNameWithCount(testLabelName, 1)).toBe(testLabelName);
+        expect(I18nUtils.getLabelNameWithCount(testLabelName, 1)).toBe(
+          testLabelName
+        );
       });
 
       it('should return singular variant if count is equal to -1', () => {
-        expect(I18nUtils.getLabelNameWithCount(testLabelName, -1)).toBe(testLabelName);
+        expect(I18nUtils.getLabelNameWithCount(testLabelName, -1)).toBe(
+          testLabelName
+        );
       });
 
       describe('given other locale', () => {
         jest.resetModules();
         jest.mock('@salesforce/i18n/locale', () => ({
-          default: 'de-DE'
+          default: 'de-DE',
         }));
         const withOtherLocale = require('c/quanticUtils');
 
         it('should return label name without failing', () => {
-          expect(() => withOtherLocale.I18nUtils.getLabelNameWithCount(testLabelName, 2)).not.toThrow();
+          expect(() =>
+            withOtherLocale.I18nUtils.getLabelNameWithCount(testLabelName, 2)
+          ).not.toThrow();
         });
       });
     });
@@ -72,21 +86,77 @@ describe('c/quanticUtils', () => {
         const testString = 'this is a {{0}} string, {{1}}.';
         const test = 'test';
         const buddy = 'buddy';
-        expect(I18nUtils.format(testString, test, buddy)).toBe('this is a test string, buddy.');
+        expect(I18nUtils.format(testString, test, buddy)).toBe(
+          'this is a test string, buddy.'
+        );
       });
     });
 
     describe('escapeHTML', () => {
       it('should escape html tags', () => {
         let htmlWithTags = '<div>test string</div>';
-      
-        expect(I18nUtils.escapeHTML(htmlWithTags)).toBe('&lt;div&gt;test string&lt;/div&gt;');
+
+        expect(I18nUtils.escapeHTML(htmlWithTags)).toBe(
+          '&lt;div&gt;test string&lt;/div&gt;'
+        );
       });
 
       it('should escape html anchor tags', () => {
         let htmlWithTags = '<a src="http://www.sketchysite.com"></a>';
-      
-        expect(I18nUtils.escapeHTML(htmlWithTags)).toBe("&lt;a src=\"http://www.sketchysite.com\"&gt;&lt;/a&gt;");
+
+        expect(I18nUtils.escapeHTML(htmlWithTags)).toBe(
+          '&lt;a src="http://www.sketchysite.com"&gt;&lt;/a&gt;'
+        );
+      });
+    });
+  });
+
+  describe('copy to clipboard', () => {
+    const mockExecCommand = jest.fn(() => {});
+    const mockWriteText = jest.fn(() => new Promise(() => {}));
+
+    beforeAll(() => {
+      Object.defineProperty(document, 'execCommand', {
+        value: mockExecCommand,
+        writable: true,
+      });
+    });
+
+    afterEach(() => {
+      jest.clearAllMocks();
+    });
+
+    describe('when the Clipboard API is available', () => {
+      it('should correctly perform the copy to clipboard operation using the writeText method', () => {
+        const exampleText = 'example text';
+
+        Object.defineProperty(navigator, 'clipboard', {
+          value: {
+            ...navigator.clipboard,
+            writeText: mockWriteText,
+          },
+          writable: true,
+        });
+
+        copyToClipboard(exampleText);
+        expect(mockWriteText).toHaveBeenCalledTimes(1);
+        expect(mockWriteText.mock.calls[0][0]).toBe(exampleText);
+        expect(mockExecCommand).toHaveBeenCalledTimes(0);
+      });
+    });
+
+    describe('when the Clipboard API is not available', () => {
+      it('should correctly perform the copy to clipboard operation using the DOM method', () => {
+        const exampleText = 'example text';
+
+        Object.defineProperty(navigator, 'clipboard', {
+          value: undefined,
+          writable: true,
+        });
+
+        copyToClipboard(exampleText);
+        expect(mockWriteText).toHaveBeenCalledTimes(0);
+        expect(mockExecCommand).toHaveBeenCalledTimes(1);
       });
     });
   });