+++ packages/atomic-angular/projects/atomic-angular/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.4](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@2.1.3...@coveo/atomic-angular@2.1.4) (2022-12-21)
+
+**Note:** Version bump only for package @coveo/atomic-angular
+
+## [2.1.3](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@2.1.2...@coveo/atomic-angular@2.1.3) (2022-12-21)
+
+**Note:** Version bump only for package @coveo/atomic-angular
+
 ## [2.1.2](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@2.1.1...@coveo/atomic-angular@2.1.2) (2022-12-19)
 
 **Note:** Version bump only for package @coveo/atomic-angular