+++ packages/quantic/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.6.0](https://github.com/coveo/ui-kit/compare/@coveo/quantic@2.5.0...@coveo/quantic@2.6.0) (2022-12-21)
+
+### Bug Fixes
+
+- **quantic:** fallback copyToClipboard method added to support this operation when Locker is ON ([#2622](https://github.com/coveo/ui-kit/issues/2622)) ([5d694c2](https://github.com/coveo/ui-kit/commit/5d694c28a6acbe29dcffb1fc6a781328a5c31be6))
+
+### Features
+
+- **Quantic:** close dropdown when Tab outside and CSS ([#2627](https://github.com/coveo/ui-kit/issues/2627)) ([4472b77](https://github.com/coveo/ui-kit/commit/4472b77cdeebbb5819b77aa076b521c7bb1282ae))
+
+# [2.5.0](https://github.com/coveo/ui-kit/compare/@coveo/quantic@2.4.0...@coveo/quantic@2.5.0) (2022-12-21)
+
+### Bug Fixes
+
+- **quantic:** behaviour of the result actions on click fixed ([#2621](https://github.com/coveo/ui-kit/issues/2621)) ([3206d8b](https://github.com/coveo/ui-kit/commit/3206d8b4389b2d739c87c1a5d6be1152eb74a9c6))
+
+### Features
+
+- **quantic:** new internal property isHovered added to the QuanticResult component ([#2617](https://github.com/coveo/ui-kit/issues/2617)) ([e71c28e](https://github.com/coveo/ui-kit/commit/e71c28e190ec29a96cef19cf2794bcff9c86e440))
+
 # [2.4.0](https://github.com/coveo/ui-kit/compare/@coveo/quantic@2.3.0...@coveo/quantic@2.4.0) (2022-12-19)
 
 ### Features